Move generate non-stream to processing.rs
